+++ src/main/java/org/example/buskmate/community/repository/CommunityPostRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CommunityPostRepository extends JpaRepository<CommunityPost,Integer> {
+public interface CommunityPostRepository extends JpaRepository<CommunityPost,Long> {
 
     @Query("""
             select cp
